+++ src/imio/events/core/indexers.py
@@ -33,6 +33,8 @@ def translated_in_en(obj):
 
 @indexer(IEvent)
 def category_title(obj):
+    if obj.local_category is not None:
+        return obj.local_category
     if obj.category is not None:
         return translate_vocabulary_term(
             "imio.events.vocabulary.EventsCategories", obj.category